[Warns]: Clear vars for translators, added missing strings to translations, added connection support

+++ sophie_bot/modules/warns.py
@@ -8,6 +8,7 @@ from sophie_bot.modules.bans import ban_user
 from sophie_bot.modules.users import (get_chat_admins, get_user_and_text,
                                       is_user_admin, user_link, get_user)
 from sophie_bot.modules.language import get_string
+from sophie_bot.modules.connections import get_conn_chat
 
 from telethon import events
 from telethon.tl.custom import Button
@@ -19,10 +20,14 @@ async def warn_user(event):
     if K is False:
         await event.reply(get_string("warns", "user_no_admeme", event.chat_id))
         return
+    status, chat_id, chat_title = await get_conn_chat(
+        event.from_id, event.chat_id, admin=True, only_in_groups=True)
+    if status is False:
+        await event.reply(chat_id)
+        return
 
     user, reason = await get_user_and_text(event)
     user_id = int(user['user_id'])
-    chat_id = event.chat_id
     if user_id in WHITELISTED:
         await event.reply(get_string("warns", "usr_whitelist", event.chat_id))
         return
@@ -42,10 +47,10 @@ async def warn_user(event):
     admin_str = await user_link(admin['user_id'])
     user_str = await user_link(user['user_id'])
     textx = get_string("warns", "warn", event.chat_id)
-    text = textx.format(admin_str, user_str)
+    text = textx.format(admin=admin_str, user=user_str, chat_name=chat_title)
     if reason:
         textx = get_string("warns", "warn_rsn", event.chat_id)
-        text += textx.format(reason)
+        text += textx.format(reason=reason)
 
     old = mongodb.warns.find({
         'user_id': user_id,
@@ -57,7 +62,7 @@ async def warn_user(event):
 
     button = Button.inline("Remove warn", 'remove_warn_{}'.format(rndm))
 
-    warn_limit = mongodb.warnlimit.find_one({'chat_id': event.chat_id})
+    warn_limit = mongodb.warnlimit.find_one({'chat_id': chat_id})
 
     if not warn_limit:
         warn_limit = 3
@@ -67,15 +72,15 @@ async def warn_user(event):
     if h >= warn_limit:
         if await ban_user(event, user_id, chat_id, None) is False:
             return
-        textx = get_string("warns", "warn_bun", event.chat_id)
-        text += textx.format(user_str)
+        textx = get_string("warns", "warn_bun", event.chat_id).format(user=user_str)
+        text += textx
         mongodb.warns.delete_many({
             'user_id': user_id,
             'group_id': chat_id
         })
     else:
         textx = get_string("warns", "warn_num", event.chat_id)
-        text += textx.format(h, warn_limit)
+        text += textx.format(curr_warns=h, max_warns=warn_limit)
 
     await event.reply(text, buttons=button, link_preview=False)
 
@@ -94,11 +99,17 @@ async def remove_warn(event):
         mongodb.notes.delete_one({'_id': warn['_id']})
     user_str = await user_link(user_id)
     textx = get_string("warns", "rmv_sfl", event.chat_id)
-    await event.edit(textx.format(user_str), link_preview=False)
+    await event.edit(textx.format(admin=user_str), link_preview=False)
 
 
 @command("warns", arg=True)
 async def user_warns(event):
+    status, chat_id, chat_title = await get_conn_chat(
+        event.from_id, event.chat_id, admin=True, only_in_groups=True)
+    if status is False:
+        await event.reply(chat_id)
+        return
+
     user, reason = await get_user_and_text(event)
     if not user:
         return
@@ -110,12 +121,11 @@ async def user_warns(event):
         return
     warns = mongodb.warns.find({
         'user_id': user_id,
-        'group_id': event.chat_id
+        'group_id': chat_id
     })
     user_str = await user_link(user_id)
-    chat_title = mongodb.chat_list.find_one({
-        'chat_id': event.chat_id})['chat_title']
-    text = "{}'s **warnings:**\n".format(user_str)
+    text = get_string("warns", "warn_list_head", event.chat_id).format(
+        admin=user_str, chat_name=chat_title)
     H = 0
     for warn in warns:
         H += 1
@@ -124,31 +134,41 @@ async def user_warns(event):
             rsn = "No reason"
         text += "{}: `{}`\n".format(H, rsn)
     if H == 0:
-        await event.reply("{} hasn't been warned in **{}** before!".format(
-            user_str, chat_title))
+        await event.reply(get_string("warns", "user_hasnt_warned", event.chat_id).format(
+            user=user_str, chat_name=chat_title))
         return
     await event.reply(text)
 
 
 @command("warnlimit", arg=True)
 async def warnlimit(event):
+    K = await is_user_admin(event.chat_id, event.from_id)
+    if K is False:
+        await event.reply(get_string("warns", "user_no_admeme", event.chat_id))
+        return
+    status, chat_id, chat_title = await get_conn_chat(
+        event.from_id, event.chat_id, admin=True, only_in_groups=True)
+    if status is False:
+        await event.reply(chat_id)
+        return
     arg = event.pattern_match.group(2)
-    old = mongodb.warnlimit.find_one({'chat_id': event.chat_id})
+    old = mongodb.warnlimit.find_one({'chat_id': chat_id})
     if not arg:
         if old:
             num = old['num']
         else:
             num = 3
-        await event.reply("Warn limit is currently: `{}`".format(num))
+        await event.reply(get_string("warns", "warn_limit", event.chat_id).format(
+            chat_name=chat_title, num=num))
     else:
         if old:
             mongodb.warnlimit.delete_one({'_id': old['_id']})
         num = int(arg)
         mongodb.warnlimit.insert_one({
-            'chat_id': event.chat_id,
+            'chat_id': chat_id,
             'num': num
         })
-        await event.reply("Warn limit has been updated to {}!".format(num))
+        await event.reply(get_string("warns", "warn_limit_upd", event.chat_id).format(num))
 
 
 @command("resetwarns", arg=True)
@@ -157,22 +177,26 @@ async def resetwarns(event):
     if K is False:
         await event.reply(get_string("warns", "user_no_admeme", event.chat_id))
         return
+    status, chat_id, chat_title = await get_conn_chat(
+        event.from_id, event.chat_id, admin=True, only_in_groups=True)
+    if status is False:
+        await event.reply(chat_id)
+        return
 
     user = await get_user(event)
     user_id = int(user['user_id'])
-    chat_id = event.chat_id
     admin = event.from_id
-    admin_str = await user_link(admin['user_id'])
+    admin_str = await user_link(admin)
     user_str = await user_link(user_id)
     chack = mongodb.warns.find({'group_id': chat_id, 'user_id': user_id})
 
     if chack:
         mongodb.warns.delete_many({'group_id': chat_id, 'user_id': user_id})
         text = get_string("warns", "purged_warns", event.chat_id)
-        await event.reply(text.format(admin_str, user_str))
+        await event.reply(text.format(admin=admin_str, user=user_str))
     else:
         text = get_string("warns", "usr_no_wrn", event.chat_id)
-        await event.reply(text.format(user_str))
+        await event.reply(text.format(user=user_str))
 
 
 def randomString(stringLength):
